+++ tensorflow/lite/experimental/micro/micro_interpreter.h
@@ -57,12 +57,12 @@ class MicroInterpreter {
   TfLiteStatus Invoke();
 
   size_t tensors_size() const { return context_.tensors_size; }
-  TfLiteTensor* tensor(int tensor_index);
+  TfLiteTensor* tensor(size_t tensor_index);
 
-  TfLiteTensor* input(int index);
+  TfLiteTensor* input(size_t index);
   size_t inputs_size() const { return subgraph_->inputs()->Length(); }
 
-  TfLiteTensor* output(int index);
+  TfLiteTensor* output(size_t index);
   size_t outputs_size() const { return subgraph_->outputs()->Length(); }
 
   TfLiteStatus initialization_status() const { return initialization_status_; }
